% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_locations.R
\name{attach_locations}
\alias{attach_locations}
\title{Add bed movement data to an extracted table}
\usage{
attach_locations(connection, ltb)
}
\arguments{
\item{connection}{a EMAP database connection}

\item{ltb}{a time series table produced from \code{\link{extract}}}
}
\value{
a time series table with location data appended
}
\description{
Takes a table created by \code{\link{extract}} and adds bed movement data.
Please be aware, this may duplicate rows, if for example a patient moves
several beds in the same hour.
}
