% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_timevarying.R
\name{expand_missing}
\alias{expand_missing}
\title{Fill in 2d Table to make a Sparse Table}
\usage{
expand_missing(df)
}
\arguments{
\item{df}{a table returned from \code{\link{extract_timevarying}}}
}
\value{
a sparse table
}
\description{
\code{\link{extract_timevarying}} returns a non-sparse table (i.e. rows/hours
with no recorded information for a patient are not presented in the table)
This function serves to expand the table and fill missing rows with NAs. This
is useful when working with packages that expect a regular cadence to the
table.
}
