% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{process_all}
\alias{process_all}
\title{Process all concepts within a group of visit_occurrences}
\usage{
process_all(
  visit_id,
  dataitems,
  visit_table,
  cadance,
  coalesce_rows,
  dttmstamp
)
}
\arguments{
\item{visit_id}{the visit_occurence_id to process}

\item{dataitems}{the table to process}

\item{visit_table}{a visit table constaining visit_start_datetime}

\item{cadance}{the base cadance}

\item{coalesce_rows}{a summary function}
}
\description{
Process all concepts within a group of visit_occurrences
}
