% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{process_item_difftime}
\alias{process_item_difftime}
\title{Process a specific concept_id}
\usage{
process_item_difftime(dataitem, var_name, start_time, cadance, coalesce_rows)
}
\arguments{
\item{dataitem}{base table}

\item{var_name}{concept_id}

\item{start_time}{start_datetime to zero from}

\item{cadance}{base time cadance}

\item{coalesce_rows}{summary function}
}
\description{
Process a concept_id in relationship to the time of hospital admission.
}
