% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{regularise}
\alias{regularise}
\title{Regularise an extracted table to a regular time cadance}
\usage{
regularise(df, cadance = 1)
}
\arguments{
\item{df}{a dense time series table produced from \code{\link{extract}}}

\item{cadance}{the cadance by which you want to expand the table
(default = 1 hour)}
}
\value{
a sparse time series table
}
\description{
The \code{\link{extract}} function returns a non-sparse table (i.e.
rows/hours with no recorded information for a patient are not presented in
the table) This function serves to expand the table and fill missing rows
with NAs. This is useful when working with most time-series aware stats
packages that expect a regular cadance to the table.
}
