% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_dummy_db}
\alias{setup_dummy_db}
\title{Setip Dummy DB}
\usage{
setup_dummy_db()
}
\value{
an SQLite in-memory database
}
\description{
Creates an in-memory SQLite database for testing purposes
}
\examples{
con <- setup_dummy_db()
res <- DBI::dbSendQuery(con, "SELECT * FROM episodes LIMIT 10")
DBI::dbFetch(res)
DBI::dbClearResult(res)
DBI::dbDisconnect(con)
}
