% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-wranglEHR.R
\docType{package}
\name{wranglEHR}
\alias{wranglEHR}
\title{wranglEHR}
\description{
wranglEHR is a data wrangling tool for EMAP. It is designed to run against
OMOP CDM 5.3.1. Please see the `R` vignettes for further details on how to
use the package to perform the most common tasks.
}
