% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_timevarying.R
\name{process_all}
\alias{process_all}
\title{Process all events}
\usage{
process_all(epi_id, events, metadata, cadence, params, time_boundaries)
}
\arguments{
\item{epi_id}{an integer vector of episode_ids to process}

\item{events}{a collected CC-HIC events table in the EAV style}

\item{metadata}{a collected CC-HIC metadata table}

\item{cadence}{baseline time cadence for extraction}

\item{params}{a table containing exactly 3 columns called: code_name,
short_name and coalesce_rows. The coalesce rows column should contain a
summary function which is used when duplicate row level data is present.}

\item{time_boundaries}{time boundaries of the extraction with respect to ICU
admission times}
}
\description{
Process all events
}
