% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_timevarying.R
\name{process_item}
\alias{process_item}
\title{Process single item}
\usage{
process_item(df, var_name, metadata, start_time, cadence, params)
}
\arguments{
\item{df}{a data table with column names as for the EAV style database tables
employed in CC-HIC. The table must contain data only for one dataitem (e.g.
have only a single unique `code_name`) and for a single episode (e.g. have
only a single unique `episode_id`)}

\item{var_name}{character string of length 1 containing the code_name to be
processed}

\item{metadata}{a table containing metadata. Can be found within the package
in the `.variables` object}

\item{start_time}{datetime of the episode start}

\item{cadence}{numeric scalar indicating the table base resolution}

\item{params}{a table containing exactly 3 columns called: code_name,
short_name and coalesce_rows. The coalesce rows column should contain a
summary function which is used when duplicate row level data is present.}
}
\description{
Takes a table in the EAV style for a single patient and a single data item
and transforms it into a tidy format
}
