% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_tables}
\alias{retrieve_tables}
\title{Retrieve DB Tables}
\usage{
retrieve_tables(connection)
}
\arguments{
\item{connection}{a DBI database object}
}
\value{
a list with pointers to DB tables
}
\description{
Retrieve DB Tables
}
